refactor: move populators from classes to root level and improve block generation
