+++ src/main/java/zone/vao/nexoAddon/NexoAddon.java
@@ -23,13 +23,13 @@ import org.bukkit.scheduler.BukkitTask;
 import org.jetbrains.annotations.NotNull;
 import zone.vao.nexoAddon.classes.Components;
 import zone.vao.nexoAddon.classes.Mechanics;
-import zone.vao.nexoAddon.classes.populators.CustomChunkGenerator;
-import zone.vao.nexoAddon.classes.populators.orePopulator.CustomOrePopulator;
-import zone.vao.nexoAddon.classes.populators.orePopulator.Ore;
-import zone.vao.nexoAddon.classes.populators.orePopulator.OrePopulator;
-import zone.vao.nexoAddon.classes.populators.treePopulator.CustomTree;
-import zone.vao.nexoAddon.classes.populators.treePopulator.CustomTreePopulator;
-import zone.vao.nexoAddon.classes.populators.treePopulator.TreePopulator;
+import zone.vao.nexoAddon.populators.CustomChunkGenerator;
+import zone.vao.nexoAddon.populators.orePopulator.CustomOrePopulator;
+import zone.vao.nexoAddon.populators.orePopulator.Ore;
+import zone.vao.nexoAddon.populators.orePopulator.OrePopulator;
+import zone.vao.nexoAddon.populators.treePopulator.CustomTree;
+import zone.vao.nexoAddon.populators.treePopulator.CustomTreePopulator;
+import zone.vao.nexoAddon.populators.treePopulator.TreePopulator;
 import zone.vao.nexoAddon.commands.NexoAddonCommand;
 import zone.vao.nexoAddon.events.*;
 import zone.vao.nexoAddon.events.blocks.BlockBreakListener;
