+++ src/main/java/zone/vao/nexoAddon/populators/orePopulator/ChunkPosition.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.classes.populators.orePopulator;
+package zone.vao.nexoAddon.populators.orePopulator;
 
 import java.util.Objects;
 import java.util.UUID;
