+++ src/main/java/zone/vao/nexoAddon/populators/treePopulator/TreePopulator.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.classes.populators.treePopulator;
+package zone.vao.nexoAddon.populators.treePopulator;
 
 import lombok.Getter;
